feature/add-api-routes feat: setup SSR and supabase
