+++ .idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/next-app.iml" filepath="$PROJECT_DIR$/.idea/next-app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/nxtable.iml" filepath="$PROJECT_DIR$/.idea/nxtable.iml" />
     </modules>
   </component>
 </project>
