+++ app/layout.tsx
@@ -1,28 +1,28 @@
 import './globals.css';
-import NavLinks from '@/app/components/nav-links/nav-links';
+import NavLinks from '@/src/shared/ui/nav-links/nav-links';
 import { Lato } from 'next/font/google';
 import { ReactNode } from 'react';
 
 const lato = Lato({
-    weight: ['400', '700', '900'],
-    subsets: ['latin'],
+  weight: ['400', '700', '900'],
+  subsets: ['latin'],
 });
 
 export default function RootLayout({
-    children,
+  children,
 }: Readonly<{
-    children: ReactNode;
+  children: ReactNode;
 }>) {
-    return (
-        <html lang="en" className={lato.className}>
-            <head>
-                <title>NXTABLE</title>
-                <link rel="icon" href="/favicon/favicon.svg" sizes="any" />
-            </head>
-            <body>
-                <NavLinks />
-                {children}
-            </body>
-        </html>
-    );
+  return (
+    <html lang="en" className={lato.className}>
+      <head>
+        <title>NXTABLE</title>
+        <link rel="icon" href="/favicon/favicon.svg" sizes="any" />
+      </head>
+      <body>
+        <NavLinks />
+        {children}
+      </body>
+    </html>
+  );
 }
