+++ app/page.tsx
@@ -1,13 +1,13 @@
-import StoreProvider from '@/app/StoreProvider';
-import BasicTable from '@/app/ui/data-table/data-table';
+import TableComponent from '@/src/features/table/TableComponent';
+import StoreProvider from '@/src/shared/store/StoreProvider';
 
-export default async function Home() {
-    return (
-        <StoreProvider>
-            <div className={'flex justify-center'}>
-                <h1>{'29.07.25'}</h1>
-                <BasicTable />
-            </div>
-        </StoreProvider>
-    );
+export default async function HomePage() {
+  return (
+    <StoreProvider>
+      <div className={'flex justify-center'}>
+        <h1>{'29.07.25'}</h1>
+        <TableComponent />
+      </div>
+    </StoreProvider>
+  );
 }
