--- app/types/index.ts
@@ -1,2 +1,0 @@
-export * from './main-clients-types';
-export * from './supabase';
