--- app/types/main-clients-types.ts
@@ -1,8 +1,0 @@
-export type Clients = Client[];
-
-export interface Client {
-    id: number;
-    name: string;
-    email: string;
-    type: string;
-}
