+++ eslint.config.mjs
@@ -7,41 +7,41 @@ import { defineConfig } from 'eslint/config';
 import { FlatCompat } from '@eslint/eslintrc';
 
 const compat = new FlatCompat({
-    baseDirectory: import.meta.dirname,
-    recommendedConfig: js.configs.recommended,
+  baseDirectory: import.meta.dirname,
+  recommendedConfig: js.configs.recommended,
 });
 
 export default defineConfig([
-    ...compat.config({
-        extends: ['eslint:recommended', 'next'],
-    }),
-    {
-        ignores: ['node_modules/**', '.next/**', 'app/types/supabase.ts'],
+  ...compat.config({
+    extends: ['eslint:recommended', 'next'],
+  }),
+  {
+    ignores: ['node_modules/**', '.next/**', 'src/shared/supabase/types/supabase.ts'],
+  },
+  {
+    files: ['**/*.{js,ts,jsx,tsx}'],
+    languageOptions: {
+      parser: tseslint.parser,
+      parserOptions: {
+        ecmaVersion: 'latest',
+        sourceType: 'module',
+        ecmaFeatures: { jsx: true },
+      },
+      globals: { ...globals.browser, ...globals.node },
     },
-    {
-        files: ['**/*.{js,ts,jsx,tsx}'],
-        languageOptions: {
-            parser: tseslint.parser,
-            parserOptions: {
-                ecmaVersion: 'latest',
-                sourceType: 'module',
-                ecmaFeatures: { jsx: true },
-            },
-            globals: { ...globals.browser, ...globals.node },
-        },
-        plugins: {
-            js,
-            react: pluginReact,
-            prettier: pluginPrettier,
-        },
-        settings: {
-            react: {
-                version: 'detect',
-            },
-        },
-        rules: {
-            'react/react-in-jsx-scope': 'off',
-            'prettier/prettier': 'warn',
-        },
+    plugins: {
+      js,
+      react: pluginReact,
+      prettier: pluginPrettier,
     },
+    settings: {
+      react: {
+        version: 'detect',
+      },
+    },
+    rules: {
+      'react/react-in-jsx-scope': 'off',
+      'prettier/prettier': 'warn',
+    },
+  },
 ]);
