--- lib/data/clients/clients.ts
@@ -1,182 +1,0 @@
-export const clientsData = [
-    {
-        id: 1,
-        name: 'John Doe',
-        email: 'johndoe@example.com',
-        type: 'individual',
-    },
-    {
-        id: 2,
-        name: 'Jane Smith',
-        email: 'janesmith@example.com',
-        type: 'individual',
-    },
-    {
-        id: 3,
-        name: 'Alice Johnson',
-        email: 'alicej@example.com',
-        type: 'individual',
-    },
-    {
-        id: 4,
-        name: 'Bob Brown',
-        email: 'bobbrown@example.com',
-        type: 'individual',
-    },
-    {
-        id: 5,
-        name: 'Charlie Davis',
-        email: 'charliedavis@example.com',
-        type: 'individual',
-    },
-    {
-        id: 6,
-        name: 'Company A',
-        email: 'contact@companya.com',
-        type: 'company',
-    },
-    {
-        id: 7,
-        name: 'Company B',
-        email: 'contact@companyb.com',
-        type: 'company',
-    },
-    {
-        id: 8,
-        name: 'Company C',
-        email: 'info@companyc.com',
-        type: 'company',
-    },
-    {
-        id: 9,
-        name: 'David White',
-        email: 'davidwhite@example.com',
-        type: 'individual',
-    },
-    {
-        id: 10,
-        name: 'Eva Green',
-        email: 'evagreen@example.com',
-        type: 'individual',
-    },
-    {
-        id: 11,
-        name: 'George Harris',
-        email: 'georgeharris@example.com',
-        type: 'individual',
-    },
-    {
-        id: 12,
-        name: 'Helen Clark',
-        email: 'helenclark@example.com',
-        type: 'individual',
-    },
-    {
-        id: 13,
-        name: 'Ian King',
-        email: 'ianking@example.com',
-        type: 'individual',
-    },
-    {
-        id: 14,
-        name: 'Jill Scott',
-        email: 'jillscott@example.com',
-        type: 'individual',
-    },
-    {
-        id: 15,
-        name: 'Jack Taylor',
-        email: 'jacktaylor@example.com',
-        type: 'individual',
-    },
-    {
-        id: 16,
-        name: 'Company D',
-        email: 'contact@companyd.com',
-        type: 'company',
-    },
-    {
-        id: 17,
-        name: 'Company E',
-        email: 'contact@companye.com',
-        type: 'company',
-    },
-    {
-        id: 18,
-        name: 'Company F',
-        email: 'info@companyf.com',
-        type: 'company',
-    },
-    {
-        id: 19,
-        name: 'Katie Moore',
-        email: 'katiemoore@example.com',
-        type: 'individual',
-    },
-    {
-        id: 20,
-        name: 'Liam Nelson',
-        email: 'liamn@example.com',
-        type: 'individual',
-    },
-    {
-        id: 21,
-        name: 'Megan Clark',
-        email: 'meganclark@example.com',
-        type: 'individual',
-    },
-    {
-        id: 22,
-        name: 'Nancy Adams',
-        email: 'nancyadams@example.com',
-        type: 'individual',
-    },
-    {
-        id: 23,
-        name: 'Oliver Lewis',
-        email: 'oliverlewis@example.com',
-        type: 'individual',
-    },
-    {
-        id: 24,
-        name: 'Paul Walker',
-        email: 'paulwalker@example.com',
-        type: 'individual',
-    },
-    {
-        id: 25,
-        name: 'Rachel Young',
-        email: 'rachelyoung@example.com',
-        type: 'individual',
-    },
-    {
-        id: 26,
-        name: 'Steve King',
-        email: 'steveking@example.com',
-        type: 'individual',
-    },
-    {
-        id: 27,
-        name: 'Company G',
-        email: 'contact@companyg.com',
-        type: 'company',
-    },
-    {
-        id: 28,
-        name: 'Company H',
-        email: 'contact@companyh.com',
-        type: 'company',
-    },
-    {
-        id: 29,
-        name: 'Company I',
-        email: 'info@companyi.com',
-        type: 'company',
-    },
-    {
-        id: 30,
-        name: 'Tom Harris',
-        email: 'tomharris@example.com',
-        type: 'individual',
-    },
-];
