+++ next.config.ts
@@ -1,11 +1,11 @@
 import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
-    /* config options here */
+  /* config options here */
 };
 
 module.exports = {
-    eslint: {},
+  eslint: {},
 };
 
 export default nextConfig;
