+++ package.json
@@ -1,53 +1,53 @@
 {
-    "name": "next-app",
-    "version": "0.1.0",
-    "private": true,
-    "scripts": {
-        "dev": "next dev --turbopack",
-        "build": "next build",
-        "types": "tsc --noEmit",
-        "start": "next start",
-        "lint": "next lint",
-        "format": "prettier --write ."
-    },
-    "dependencies": {
-        "@next/eslint-plugin-next": "^15.4.4",
-        "@reduxjs/toolkit": "^2.8.2",
-        "@supabase/ssr": "^0.6.1",
-        "@supabase/supabase-js": "^2.52.1",
-        "@tanstack/react-table": "^8.21.3",
-        "class-variance-authority": "^0.7.1",
-        "clsx": "^2.1.1",
-        "eslint-plugin-react-hooks": "^5.2.0",
-        "lucide-react": "^0.525.0",
-        "next": "15.4.4",
-        "react": "19.1.0",
-        "react-dom": "19.1.0",
-        "react-redux": "^9.2.0",
-        "tailwind-merge": "^3.3.1"
-    },
-    "devDependencies": {
-        "@eslint/eslintrc": "^3",
-        "@eslint/js": "^9.31.0",
-        "@tailwindcss/postcss": "^4.1.11",
-        "@types/node": "^20",
-        "@types/react": "^19",
-        "@types/react-dom": "^19",
-        "@typescript-eslint/eslint-plugin": "^8.38.0",
-        "@typescript-eslint/parser": "^8.38.0",
-        "autoprefixer": "^10.4.21",
-        "eslint": "^9.31.0",
-        "eslint-config-next": "15.4.1",
-        "eslint-config-prettier": "^10.1.8",
-        "eslint-plugin-prettier": "^5.5.3",
-        "eslint-plugin-react": "^7.37.5",
-        "eslint-plugin-unused-imports": "^4.1.4",
-        "globals": "^16.3.0",
-        "postcss": "^8.5.6",
-        "prettier": "3.6.2",
-        "tailwindcss": "^4.1.11",
-        "tw-animate-css": "^1.3.5",
-        "typescript": "^5",
-        "typescript-eslint": "^8.37.0"
-    }
+  "name": "nxtable",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "dev": "next dev --turbopack",
+    "build": "next build",
+    "types": "tsc --noEmit",
+    "start": "next start",
+    "lint": "next lint",
+    "format": "prettier --write ."
+  },
+  "dependencies": {
+    "@next/eslint-plugin-next": "^15.5.4",
+    "@reduxjs/toolkit": "^2.9.0",
+    "@supabase/ssr": "^0.6.1",
+    "@supabase/supabase-js": "^2.58.0",
+    "@tanstack/react-table": "^8.21.3",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "lucide-react": "^0.525.0",
+    "next": "15.4.7",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-redux": "^9.2.0",
+    "tailwind-merge": "^3.3.1"
+  },
+  "devDependencies": {
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.36.0",
+    "@tailwindcss/postcss": "^4.1.14",
+    "@types/node": "^20.19.19",
+    "@types/react": "^19.2.0",
+    "@types/react-dom": "^19.2.0",
+    "@typescript-eslint/eslint-plugin": "^8.45.0",
+    "@typescript-eslint/parser": "^8.45.0",
+    "autoprefixer": "^10.4.21",
+    "eslint": "^9.36.0",
+    "eslint-config-next": "15.4.1",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.4",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-unused-imports": "^4.2.0",
+    "globals": "^16.4.0",
+    "postcss": "^8.5.6",
+    "prettier": "3.6.2",
+    "tailwindcss": "^4.1.14",
+    "tw-animate-css": "^1.4.0",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.45.0"
+  }
 }
