+++ postcss.config.mjs
@@ -1,3 +1,3 @@
 export default {
-    plugins: ['@tailwindcss/postcss'],
+  plugins: ['@tailwindcss/postcss'],
 };
