+++ prettier.config.cjs
@@ -1,6 +1,6 @@
 module.exports = {
-    semi: true,
-    singleQuote: true,
-    trailingComma: 'all',
-    tabWidth: 4,
+  semi: true,
+  singleQuote: true,
+  trailingComma: 'all',
+  tabWidth: 2,
 };
