+++ src/entities/clients/types/types.ts
@@ -1,0 +1,8 @@
+export type Clients = Client[];
+
+export interface Client {
+  id: number;
+  name: string;
+  email: string;
+  type: string;
+}
