+++ src/features/table/types/index.ts
@@ -1,0 +1,1 @@
+export * from './column';
