+++ src/shared/store/utils/hooks.ts
@@ -1,4 +1,4 @@
-import { AppDispatch, AppStore, RootState } from '@/lib/store';
+import { AppDispatch, AppStore, RootState } from '@/src/shared/store/store';
 import { useDispatch, useSelector, useStore } from 'react-redux';
 
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
