--- utils/supabase/client.ts
@@ -1,6 +1,0 @@
-import { createClient } from '@supabase/supabase-js';
-
-export const supabase = createClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL as string,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY as string,
-);
